Generify ActionLineParser into LineParser<Action>
