+++ tichu-cli/src/main/java/net/incongru/tichu/simu/parse/LineParser.java
@@ -1,0 +1,7 @@
+package net.incongru.tichu.simu.parse;
+
+interface LineParser<T> {
+    boolean accept(TokenisedLine t);
+
+    T parse(TokenisedLine t);
+}
